updated the doxygen comments and cleanup

+++ utilities/test_suite/HIP/Tensor_image_hip.cpp
@@ -1698,7 +1698,7 @@ int main(int argc, char **argv)
                     std::ofstream refFile;
                     refFile.open(func + ".csv");
                     for (int i = 0; i < oBufferSize; i++)
-                        refFile << static_cast<float>(*(outputu8 + i)) << ",";
+                        refFile << static_cast<int>(*(outputu8 + i)) << ",";
                     refFile.close();
                 }
